+++ frontend/src/components/Questions/editors/index.ts
@@ -1,0 +1,8 @@
+export * from './QuestionEditor'
+export * from './MCQEditor'
+export * from './TrueFalseEditor'
+export * from './ShortAnswerEditor'
+export * from './EssayEditor'
+export * from './FillInBlankEditor'
+export * from './UnsupportedEditor'
+export * from './ErrorEditor'
